server_bios_agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
gradio==4.12.0
requests==2.31.0